refactor: rename `pasuableWatch` to `watchPasuable` with back alias
