+++ packages/shared/watchPausable/demo.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
 import { ref } from 'vue-demi'
 import { onStartTyping } from '@vueuse/core'
-import { pausableWatch } from '.'
+import { watchPausable } from '.'
 
 const input = ref<HTMLInputElement | null>()
 const log = ref('')
 
 const source = ref('')
 
-const watcher = pausableWatch(
+const watcher = watchPausable(
   source,
   v => (log.value += `Changed to "${v}"\n`),
 )
